Adjust control logistics, SA previews, etc...

+++ Assembly-CSharp/Global/POINTS.cs
@@ -3,12 +3,8 @@
 public class POINTS
 {
     public UInt32 hp;
-
     public UInt32 mp;
-
     public Int16 at;
-
     public SByte at_coef;
-
-    public Byte capa;
+    public UInt32 capa;
 }
